requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.11
pyjwt==2.11.0
passlib==1.7.4
bcrypt==4.0.1
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-auth==2.25.2
python-jose==3.3.0